+++ src/components/literatureReview/LiteratureReview.js
@@ -1,23 +1,17 @@
 import React from 'react';
-import './LiteratureReview.css';
+import PageContainer, { Section } from '../pageContainer/PageContainer';
 
 const LiteratureReview = () => (
-  <div className="LiteratureReview">
-    <div className="titleContainer">
-      <p className="literatureTitle">Literature Review</p>
-    </div>
-
-    <div className="overviewContainer">
-      <p className="overview">
-        {' '}
+  <PageContainer title="Literature Review">
+    <Section>
+      <p style={{ fontSize: '22px' }}>
         <span style={{ fontWeight: 'bold' }}>Problem Overview</span>: How has COVID-19
         impacted the sustainable energy transition in North America
       </p>
-    </div>
+    </Section>
 
-    <div className="overviewContainer">
-      <p className="review">
-        {' '}
+    <Section>
+      <p style={{ fontSize: '25px' }}>
         The COVID-19 pandemic has impacted almost every aspect of our lives. The pandemic
         has has caused delays in many renewable energy manufacturing facilities, and
         supply chains. In return many sustainable energy companies are slowing down which
@@ -27,18 +21,17 @@ const LiteratureReview = () => (
         early to tell how bad these effects will be in the long run, but sustainable
         energy policies need to be reconsidered.
       </p>
-    </div>
+    </Section>
 
-    <div className="overviewContainer">
-      <p className="cite">
-        {' '}
+    <Section>
+      <p>
         <span style={{ fontWeight: 'bold' }}>Citation</span>: S. Ehsan Hosseini, An
         Outlook on the Global Development of Renewable and Sustainable Energy at the Time
         of Covid-19, Energy Research & Social Science (2020), doi: https://doi.org/
         10.1016/j.erss.2020.101633
       </p>
-    </div>
-  </div>
+    </Section>
+  </PageContainer>
 );
 
 export default LiteratureReview;
