+++ src/components/pageContainer/PageContainer.js
@@ -4,7 +4,7 @@ import './PageContainer.css';
 
 const Section = ({ className = '', title = '', children = [] }) => (
   <div className={'section-container ' + className}>
-    <h2 className="section-title">{title}</h2>
+    {title.length > 0 ? <h2 className="section-title">{title}</h2> : ''}
     {children || ''}
   </div>
 );
@@ -12,6 +12,15 @@ const Section = ({ className = '', title = '', children = [] }) => (
 const Image = ({ className = '', alt = 'No alt text provided.', src = '' }) => (
   <div className={'image-container ' + className}>
     <img src={src} alt={alt} />
+  </div>
+);
+
+const Reference = ({ id, text }) => (
+  <div style={{ paddingBottom: '30px', paddingLeft: '30px' }}>
+    <span style={{ marginLeft: '-30px' }}>[{id}]</span>
+    <span style={{ paddingLeft: '30px' }}>
+      <em>{text}</em>
+    </span>
   </div>
 );
 
@@ -37,4 +46,4 @@ const PageContainer = ({ className = '', title = '', children = [] }) => {
   );
 };
 
-export { PageContainer as default, Section, Image };
+export { PageContainer as default, Section, Image, Reference };
